+++ SentryTestUtils/ClearTestState.swift
@@ -5,6 +5,16 @@ public func clearTestState() {
     TestCleanup.clearTestState()
 }
 
+public func resetUserDefaults() {
+    if let appDomain = Bundle.main.bundleIdentifier {
+        UserDefaults.standard.removePersistentDomain(forName: appDomain)
+        // Although the Apple docs state this shouldn't be used we need it
+        // to avoid race conditions in tests for UserDefaults. Not calling
+        // this can lead to flaky tests.
+        UserDefaults.standard.synchronize()
+    }
+}
+
 @objcMembers
 class TestCleanup: NSObject {
     static func clearTestState() {