+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ## Unreleased
 
-### Fixes
+## Feature
 
+- Add option to report uncaught NSExceptions on macOS (#4471)
 - Build visionOS project with static Sentry SDK (#4462)
 - Too many navigation breadcrumbs for Session Replay (#4480)
+- Capture all touches with session replay (#4477)
 
 ### Improvements